Fix bootstrap incompatibility after multipass 1.10.0
With multipass 1.10.0 the behavior of multipass exec was changed from
always using the home folder to instead use current folder if said
folder is mounted on the remote filesystem.

If your current folder was mounted in the multipass VM (by default home
folder is), docker-multipass create would fail.

This fixes said bug.